[ any & all ] list predicates.
We need to be careful when thinking about higher predicates, especially with
respect to the indexed predicate's 'polarity'.
Where polarity indicates if we are using the positive or negative truth.

Take, for example, `ANY` and `ALL`, both of which use `All` and `Any`
quantifiers in their construction.

For a given predicate `p(+,-)`, we can say that:

+ `prf : All p+ p- xs` := [] | p+ x :: prf`

  That is `All` is positive in all constructors.

+ `prf : Any p+ p- xs := p- x | p+ x :: prf`

  That is `Any` is negative in the `Here` case (i.e. we found the counter example), and positive in the `There` case.

This does not mean that `All p+ p- xs <=> Any p+ p- xs` as what is
positive and negative depends on what is means for `p` to hold in the quantifier.

That is, for a given predicate `p(+,-)`, we need to frame `ALL` and
`ANY` as:

+ `ALL p xs := (All p+ p- xs, Any p+ p-)`

  The predicate holds unless we find a negative case.

+ `ANY p xs := (Any p- p+ xs, All p- p+)`

  The predicate fails until we find a positive case.