[ start ] elaborating the STLC.
This is going to be hard.

Positive `ELEM` requires proof of false in `There` case, this makes creating it harder when the typing context is just types...

Need to think of alternative ways to construct an `Elem` but interfaces are making it harder to resolve helper functions to turn `DECEQpos` into `Equals`.